Done some refactoring inside the inWork package. Working on the play state
