--- src/main/java/inf112/skeleton/app/InWork/RoboRallyBoard.java
@@ -1,55 +1,0 @@
-package inf112.skeleton.app.InWork;
-
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.maps.tiled.TiledMap;
-import com.badlogic.gdx.maps.tiled.TmxMapLoader;
-import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
-
-public class RoboRallyBoard implements Screen {
-
-    private TiledMap board;
-    private OrthogonalTiledMapRenderer renderer;
-    private OrthographicCamera camera;
-
-    @Override
-    public void show() {
-        camera = new OrthographicCamera();
-        board = new TmxMapLoader().load("assets/RoboRallyMap.tmx"); // Loads the .tmx file, which is the board
-        renderer = new OrthogonalTiledMapRenderer(board);
-    }
-
-    @Override
-    public void render(float delta) {
-    Gdx.gl.glClearColor(0,0,0,1);
-    Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-    renderer.render();
-    }
-
-    @Override
-    public void resize(int width, int height) {
-    }
-
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-    dispose();
-    }
-
-    @Override
-    public void dispose() {
-    board.dispose();
-    renderer.dispose();
-    }
-}
