--- src/main/java/inf112/skeleton/app/InWork/StateManager.java
@@ -1,4 +1,0 @@
-package inf112.skeleton.app.InWork;
-
-public class StateManager {
-}
