+++ src/main/java/inf112/skeleton/app/Main.java
@@ -2,7 +2,6 @@ package inf112.skeleton.app;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import inf112.skeleton.app.InWork.NewRoboRally;
 
 public class Main {
