+++ src/main/java/inf112/skeleton/app/inWork/NewRoboRally.java
@@ -1,34 +1,33 @@
-package inf112.skeleton.app.InWork;
+package inf112.skeleton.app.inWork;
 
 import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import inf112.skeleton.app.inWork.state.StartState;
+import inf112.skeleton.app.inWork.state.StateManager;
 
 
 public class NewRoboRally extends ApplicationAdapter {
     public static final String TITLE = "RoboRally";
     public static final int WIDTH = 480;
     public static final int HEIGHT = 800;
-
-    SpriteBatch batch;
-    Texture greenrobot;
+    private StateManager statemanager;
+    private SpriteBatch batch;
 
     @Override
     public void create() {
+        statemanager = new StateManager();
         batch = new SpriteBatch();
-        greenrobot = new Texture("assets/GreenRobot.png");
+        Gdx.gl.glClearColor(1,0,0,1);
+        statemanager.push(new StartState(statemanager));
     }
 
     @Override
     public void render() {
-        Gdx.gl.glClearColor(1,0,0,1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-
-        batch.begin();
-        batch.draw(greenrobot,0,0);
-        batch.end();
+        statemanager.update(Gdx.graphics.getDeltaTime());
+        statemanager.render(batch);
     }
 
 }
