+++ src/main/java/inf112/skeleton/app/tiles/Floor.java
@@ -1,10 +1,15 @@
-package inf112.skeleton.app.Tiles;
+package inf112.skeleton.app.tiles;
 
-public class Wall implements ITile {
+public class Floor implements ITile {
 
-    String tileType = "wall";
-    int tileID = 2;
-    int tileWeight = 2;
+    String tileType = "floor";
+    int tileID;
+    int tileWeight;
+
+    public Floor() {
+        this.tileID = 1;
+        this.tileWeight = 1;
+    }
 
     @Override
     public String getTileType() {
